+++ src/main/java/com/crud/crudsimple/models/Cidade.java
@@ -25,7 +25,7 @@ public class Cidade {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "cid_id", unique = true, nullable = false, updatable = false)
-	private int idCidade;
+	private Long idCidade;
 
 	@Column(name = "cid_nome", nullable = false, length = 100)
 	@NotNull (groups = {CreateCidade.class, UpdateCidade.class})
